feat(management): init

+++ myopia-business/bootstrap/src/main/resources/db/migration/V20220401__V2.4.0-common-disease.sql
@@ -1,6 +1,6 @@
 -- sql 加上注释说明描述
 alter table m_stat_rescreen
-    add type int default 0 null comment '0-视力筛查 1-常见病' after school_id;
+    add type tinyint(1) default 0 null comment '0-视力筛查 1-常见病' after school_id;
 
 alter table m_stat_rescreen
     add physique_rescreen_num int null comment '体格复查人数';
